+++ react-tutorial/src/App.js
@@ -1,4 +1,4 @@
-import Board from "./components/board";
+import Board from "./Components/Board";
 
 export default function App() {
   return <Board />;
