+++ react-tutorial/src/components/board/styles.css
@@ -1,0 +1,9 @@
+.status {
+  margin-bottom: 10px;
+}
+
+.board-row:after {
+  clear: both;
+  content: "";
+  display: table;
+}
