--- react-tutorial/src/components/square/Square.js
@@ -1,8 +1,0 @@
-function Square({ value, onSquareClick }) {
-  return (
-    <button className="square" onClick={onSquareClick}>
-      {value}
-    </button>
-  );
-}
-export default Square;
