+++ react-tutorial/src/components/square/index.js
@@ -1,1 +1,10 @@
-export { default } from "./Square";
+import "./styles.css";
+
+function Square({ value, onSquareClick }) {
+  return (
+    <button className="square" onClick={onSquareClick}>
+      {value}
+    </button>
+  );
+}
+export default Square;
