+++ react-tutorial/src/index.js
@@ -1,6 +1,5 @@
 import React, { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./styles.css";
 
 import App from "./App";
